feat: add matomo on /app (#935)
